fix(ai-playground): repoint the two generic components moved out of ai-skill/
`BsTimestamp` and `BsLatency` now live in `components/` (PR1 rebase): the
playground was importing them through the ai-skill module's folder, which is
exactly why they were moved.
